refactor: extract SplitExpenseItems.ts from Split.ts
Move 12 split-expense-item helper functions out of the now-large Split.ts
into a dedicated module. This is a pure move — function bodies are
byte-identical to their previous location.

Functions moved:
- getDistanceMerchantFromDistance
- updateSplitExpenseDistanceFromAmount
- initSplitExpenseItemData
- initDraftSplitExpenseDataForEdit
- redistributeSplitExpenseAmounts
- addSplitExpenseField
- evenlyDistributeSplitExpenseAmounts
- resetSplitExpensesByDateRange
- removeSplitExpenseField
- updateSplitExpenseField
- updateSplitExpenseAmountField
- clearSplitTransactionDraftErrors

Consumers updated to import from the new location:
- src/hooks/useDeleteTransactions.ts
- src/libs/actions/SplitExpenses.ts
- src/pages/iou/SplitExpenseCreateDateRagePage.tsx
- src/pages/iou/SplitExpenseEditPage.tsx
- src/pages/iou/SplitExpensePage.tsx
- tests/actions/IOUTest/SplitTest.ts
- tests/unit/SplitExpenseAutoAdjustmentTest.ts

Follow-up to PR #87957. Next PR will extract updateSplitTransactions.
