+++ src/hooks/useDeleteTransactions.ts
@@ -5,7 +5,8 @@ import {useSearchStateContext} from '@components/Search/SearchContext';
 import {getIOURequestPolicyID} from '@libs/actions/IOU';
 import {deleteMoneyRequest} from '@libs/actions/IOU/DeleteMoneyRequest';
 import {getIOUActionForTransactions} from '@libs/actions/IOU/Duplicate';
-import {initSplitExpenseItemData, updateSplitTransactions} from '@libs/actions/IOU/Split';
+import {updateSplitTransactions} from '@libs/actions/IOU/Split';
+import {initSplitExpenseItemData} from '@libs/actions/IOU/SplitExpenseItems';
 import getNonEmptyStringOnyxID from '@libs/getNonEmptyStringOnyxID';
 import {getOriginalMessage, isMoneyRequestAction} from '@libs/ReportActionsUtils';
 import {getActiveGroupSearchHashes} from '@libs/SearchUIUtils';
